tin_services/requirements.txt
APScheduler==3.10.4
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
docker==7.1.0
Flask==3.1.0
Flask-Cors==3.0.10
flask-marshmallow==1.3.0
flask-swagger-ui==4.11.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
marshmallow==3.26.0
mistune==3.0.2
nodeenv==1.9.1
packaging==24.2
pymongo==4.10.1
python-vagrant==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
